+++ packages/types/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @moonbeam-network/xcm-types
 
+## 4.2.3
+
+### Patch Changes
+
+- [#572](https://github.com/moonbeam-foundation/xcm-sdk/pull/572) [`b6ad5e8`](https://github.com/moonbeam-foundation/xcm-sdk/commit/b6ad5e8059fccac6513dd1de0484e7f6a205cb62) Thanks [@mmaurello](https://github.com/mmaurello)! - Update dependencies
+
+- Updated dependencies [[`b6ad5e8`](https://github.com/moonbeam-foundation/xcm-sdk/commit/b6ad5e8059fccac6513dd1de0484e7f6a205cb62)]:
+  - @moonbeam-network/xcm-utils@4.1.3
+
 ## 4.2.2
 
 ### Patch Changes